Add App page fully redesigned
