+++ README.md
@@ -4,7 +4,7 @@
 
 ObtainX is a fork of Obtainium. These are the extra features you get in this fork:
 
-- **📦 Installer choice** - Added a new **Legacy** install path. It sends APKs to any installer you choose (e.g. InstallerX, App Manager). Useful when you cannot grant "install unknown apps" to normal apps (e.g. when _Advanced Protection_ is enabled) but a privileged installer can still do the job.
+- **📦 Installer choice** - Added a **Third-Party** install path. It sends APKs to any installer you choose (e.g. InstallerX, App Manager). Useful when you cannot grant "install unknown apps" to normal apps (e.g. when _Advanced Protection_ is enabled) but a privileged installer can still do the job.
 
 - **🎨 Material 3 Expressive Makeover** - Different parts of the app has been overhauled with Material 3 Expressive design, with grouping in cards, slide up panels, fluid animations, expressive buttons, auto-hide menu bars, and visual consistency tweaks.
