+++ assets/translations/en.json
@@ -54,6 +54,10 @@
     "appAlreadyAdded": "App already added",
     "alreadyUpToDateQuestion": "App already up to date?",
     "addApp": "Add app",
+    "addByUrl": "URL",
+    "addBySearch": "Search",
+    "addFromDevice": "Device",
+    "searchStores": "Search stores",
     "appSourceURL": "App source URL",
     "error": "Error",
     "add": "Add",
@@ -378,10 +382,10 @@
     "installerMode": "Installer mode",
     "installerModeStock": "Stock",
     "installerModeShizuku": "Shizuku",
-    "installerModeLegacy": "Legacy",
-    "legacyInstallerSelect": "Select installer app",
-    "legacyInstallerNoneSelected": "No installer selected",
-    "legacyInstallerNotSelected": "Please select a legacy installer app in Settings first",
+    "installerModeThirdParty": "Third-party",
+    "thirdPartyInstallerSelect": "Select installer app",
+    "thirdPartyInstallerNoneSelected": "No installer selected",
+    "thirdPartyInstallerNotSelected": "Please select a third-party installer app in Settings first",
     "useSystemFont": "Use the system font",
     "useVersionCodeAsOSVersion": "Use app versionCode as OS-detected version",
     "requestHeader": "Request header",
@@ -524,7 +528,6 @@
     "excludeAlreadyTrackedApps": "Exclude apps that are already tracked",
     "deleteBulkScanHistory": "Delete scan data of previous searches",
     "deleteBulkScanHistorySubtitle": "Clears saved store lookups when you tap Find apps. Does not remove apps from ObtainX.",
-    "bulkAddNextScreenSelectAppsNote": "On the next screen, choose which apps to include. Nothing is scanned until you tap Find apps.",
     "bulkScanProgressXY": "{}/{} searched",
     "cancelBulkScan": "Cancel search",
     "bulkScanCancelled": "Cancelled",
@@ -532,7 +535,7 @@
     "bulkAddChipSystemApps": "System Apps",
     "bulkAddChipAllApps": "All Apps",
     "storesToSearch": "Stores to search",
-    "bulkScanCacheNote": "Lookups are saved per app and per store. When you run a new scan, only stores you turn on here are queried. Cached rows from earlier scans are reused for those same stores (APKMirror stays instant if it was already checked). A store you never checked before is always fetched live. Example: last scan used APKMirror and APKPure; this scan uses APKMirror and F-Droid – APKMirror is read from disk, F-Droid is queried for every selected app.",
+    "bulkScanCacheNote": "Results are cached per app, per store. Each app is only re-fetched from a store if it wasn't scanned there before.",
     "selectAtLeastOneStore": "Select at least one store",
     "next": "Next",
     "selectAppsToImport": "Select apps",
