+++ lib/pages/add_app.dart
@@ -2,6 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:obtainium/components/app_page_section_title.dart';
+import 'package:obtainium/components/bulk_add_widget.dart';
 import 'package:obtainium/components/custom_app_bar.dart';
 import 'package:obtainium/components/generated_form.dart';
 import 'package:obtainium/components/version_regex_assist_dialog.dart';
@@ -10,8 +11,6 @@ import 'package:obtainium/custom_errors.dart';
 import 'package:obtainium/main.dart';
 import 'package:obtainium/pages/app.dart';
 import 'package:obtainium/pages/page_route_slide_up.dart';
-import 'package:obtainium/pages/bulk_add_apps.dart';
-import 'package:obtainium/pages/import_export.dart';
 import 'package:obtainium/pages/settings.dart';
 import 'package:obtainium/providers/apps_provider.dart';
 import 'package:obtainium/providers/notifications_provider.dart';
@@ -22,6 +21,8 @@ import 'package:obtainium/theme/app_page_icon_colors.dart';
 import 'package:provider/provider.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
+enum _AddMode { byUrl, search, fromDevice }
+
 class AddAppPage extends StatefulWidget {
   const AddAppPage({super.key});
 
@@ -30,11 +31,12 @@ class AddAppPage extends StatefulWidget {
 }
 
 class AddAppPageState extends State<AddAppPage> {
-  bool gettingAppInfo = false;
-  bool searching = false;
+  // ─── Mode ──────────────────────────────────────────────────────────────
+  _AddMode _mode = _AddMode.byUrl;
 
+  // ─── URL mode state ────────────────────────────────────────────────────
+  bool gettingAppInfo = false;
   String userInput = '';
-  String searchQuery = '';
   String? pickedSourceOverride;
   String? previousPickedSourceOverride;
   AppSource? pickedSource;
@@ -44,7 +46,31 @@ class AddAppPageState extends State<AddAppPage> {
   List<String> pickedCategories = [];
   SourceProvider sourceProvider = SourceProvider();
   late final TextEditingController _urlFieldController;
+
+  // ─── Device mode state ────────────────────────────────────────────────
+  final GlobalKey<BulkAddWidgetState> _bulkWidgetKey = GlobalKey();
+
+  /// Called by [HomePageState] when the user presses back while this tab is
+  /// active. Returns true if the bulk flow consumed the event (moved one step
+  /// back). Returns false so the caller falls through to normal tab navigation.
+  bool handleBack() {
+    if (_mode == _AddMode.fromDevice) {
+      return _bulkWidgetKey.currentState?.handleBack() ?? false;
+    }
+    return false;
+  }
+
+  // ─── Search mode state ─────────────────────────────────────────────────
+  bool searching = false;
+  String searchQuery = '';
+  // Searchable-source names the user has selected (null = not yet initialised)
+  Set<String>? _searchSelectedStores;
+  // Interleaved search results: key=URL/identifier, value=(sourceName, subtitleLines)
+  Map<String, MapEntry<String, List<String>>> _searchResults = {};
+  bool _searchHasSearched = false;
+  String _searchResultFilter = '';
   late final TextEditingController _searchSomeSourcesController;
+  late final TextEditingController _searchResultFilterController;
   late final FocusNode _searchSomeSourcesFocusNode;
 
   void linkFn(String input) {
@@ -64,6 +90,7 @@ class AddAppPageState extends State<AddAppPage> {
     super.initState();
     _urlFieldController = TextEditingController();
     _searchSomeSourcesController = TextEditingController();
+    _searchResultFilterController = TextEditingController();
     _searchSomeSourcesFocusNode = FocusNode();
   }
 
@@ -71,8 +98,21 @@ class AddAppPageState extends State<AddAppPage> {
   void dispose() {
     _urlFieldController.dispose();
     _searchSomeSourcesController.dispose();
+    _searchResultFilterController.dispose();
     _searchSomeSourcesFocusNode.dispose();
     super.dispose();
+  }
+
+  /// Lazily initialise [_searchSelectedStores] using persisted deselections.
+  Set<String> _getSearchSelectedStores(SettingsProvider settingsProvider) {
+    if (_searchSelectedStores == null) {
+      final deselected = settingsProvider.searchDeselected.toSet();
+      _searchSelectedStores = sourceProvider.sources
+          .where((e) => e.canSearch && !deselected.contains(e.name))
+          .map((e) => e.name)
+          .toSet();
+    }
+    return _searchSelectedStores!;
   }
 
   bool _isUrlInputValid(String value) {
@@ -149,6 +189,8 @@ class AddAppPageState extends State<AddAppPage> {
 
     bool doingSomething = gettingAppInfo || searching;
 
+    // ── Track-only / release-date confirmations (URL mode) ─────────────
+
     Future<bool> getTrackOnlyConfirmationIfNeeded(
       bool userPickedTrackOnly, {
       bool ignoreHideSetting = false,
@@ -202,6 +244,8 @@ class AddAppPageState extends State<AddAppPage> {
               ) ==
               null));
     }
+
+    // ── Add app (URL mode) ─────────────────────────────────────────────
 
     addApp({bool resetUserInputAfter = false}) async {
       setState(() {
@@ -257,8 +301,6 @@ class AddAppPageState extends State<AddAppPage> {
             throw ObtainiumError(tr('appAlreadyAdded'));
           }
           if (app.additionalSettings['trackOnly'] == true) {
-            // Installed version only comes from PackageManager when [app.id] is
-            // a real package name. Temp IDs cannot query the device.
             app.installedVersion = null;
             if (isTempId(app)) {
               app.additionalSettings['trackOnlyTemporaryPackageId'] = true;
@@ -306,6 +348,8 @@ class AddAppPageState extends State<AddAppPage> {
         });
       }
     }
+
+    // ── URL mode widgets ───────────────────────────────────────────────
 
     Widget getUrlInputRow() {
       final ColorScheme colorScheme = Theme.of(context).colorScheme;
@@ -381,178 +425,6 @@ class AddAppPageState extends State<AddAppPage> {
       );
     }
 
-    runSearch({bool filtered = true}) async {
-      _searchSomeSourcesFocusNode.unfocus();
-      FocusManager.instance.primaryFocus?.unfocus();
-      setState(() {
-        searching = true;
-      });
-      var sourceStrings = <String, List<String>>{};
-      sourceProvider.sources.where((e) => e.canSearch).forEach((s) {
-        sourceStrings[s.name] = [s.name];
-      });
-      try {
-        var searchSources =
-            await showModalBottomSheet<List<String>?>(
-              context: context,
-              isScrollControlled: true,
-              useSafeArea: true,
-              shape: const RoundedRectangleBorder(
-                borderRadius: BorderRadius.vertical(top: Radius.circular(24)),
-              ),
-              builder: (BuildContext ctx) {
-                return SelectionModal(
-                  presentAsBottomSheet: true,
-                  showFilterField: false,
-                  title: tr(
-                    'selectX',
-                    args: [plural('source', 2).toLowerCase()],
-                  ),
-                  entries: sourceStrings,
-                  selectedByDefault: true,
-                  onlyOneSelectionAllowed: false,
-                  titlesAreLinks: false,
-                  deselectThese: settingsProvider.searchDeselected,
-                );
-              },
-            ) ??
-            [];
-        if (searchSources.isNotEmpty) {
-          settingsProvider.searchDeselected = sourceStrings.keys
-              .where((s) => !searchSources.contains(s))
-              .toList();
-          List<MapEntry<String, Map<String, List<String>>>?>
-          results = (await Future.wait(
-            sourceProvider.sources
-                .where((e) => searchSources.contains(e.name))
-                .map((e) async {
-                  try {
-                    Map<String, dynamic>? querySettings = {};
-                    if (e.includeAdditionalOptsInMainSearch) {
-                      querySettings = await showDialog<Map<String, dynamic>?>(
-                        context: context,
-                        builder: (BuildContext ctx) {
-                          return GeneratedFormModal(
-                            title: tr('searchX', args: [e.name]),
-                            items: [
-                              ...e.searchQuerySettingFormItems.map((e) => [e]),
-                              [
-                                GeneratedFormTextField(
-                                  'url',
-                                  label: e.hosts.isNotEmpty
-                                      ? tr('overrideSource')
-                                      : plural('url', 1).substring(2),
-                                  autoCompleteOptions: [
-                                    ...(e.hosts.isNotEmpty ? [e.hosts[0]] : []),
-                                    ...appsProvider.apps.values
-                                        .where(
-                                          (a) =>
-                                              sourceProvider
-                                                  .getSource(
-                                                    a.app.url,
-                                                    overrideSource:
-                                                        a.app.overrideSource,
-                                                  )
-                                                  .runtimeType ==
-                                              e.runtimeType,
-                                        )
-                                        .map((a) {
-                                          var uri = Uri.parse(a.app.url);
-                                          return '${uri.origin}${uri.path}';
-                                        }),
-                                  ],
-                                  defaultValue: e.hosts.isNotEmpty
-                                      ? e.hosts[0]
-                                      : '',
-                                  required: true,
-                                ),
-                              ],
-                            ],
-                          );
-                        },
-                      );
-                      if (querySettings == null) {
-                        return null;
-                      }
-                    }
-                    return MapEntry(
-                      e.runtimeType.toString(),
-                      await e.search(searchQuery, querySettings: querySettings),
-                    );
-                  } catch (err) {
-                    if (err is! CredsNeededError) {
-                      rethrow;
-                    } else {
-                      err.unexpected = true;
-                      if (!context.mounted) return null;
-                      showError(err, context);
-                      return null;
-                    }
-                  }
-                }),
-          )).where((a) => a != null).toList();
-
-          // Interleave results instead of simple reduce
-          Map<String, MapEntry<String, List<String>>> res = {};
-          var si = 0;
-          var done = false;
-          while (!done) {
-            done = true;
-            for (var r in results) {
-              var sourceName = r!.key;
-              if (r.value.length > si) {
-                done = false;
-                var singleRes = r.value.entries.elementAt(si);
-                res[singleRes.key] = MapEntry(sourceName, singleRes.value);
-              }
-            }
-            si++;
-          }
-          if (res.isEmpty) {
-            throw ObtainiumError(tr('noResults'));
-          }
-          if (!context.mounted) return;
-          List<String>? selectedUrls = res.isEmpty
-              ? []
-              : await showModalBottomSheet<List<String>?>(
-                  context: context,
-                  isScrollControlled: true,
-                  useSafeArea: true,
-                  shape: const RoundedRectangleBorder(
-                    borderRadius: BorderRadius.vertical(top: Radius.circular(24)),
-                  ),
-                  builder: (BuildContext ctx) {
-                    return SelectionModal(
-                      presentAsBottomSheet: true,
-                      showFilterField: true,
-                      title: tr('addAppSearchResultsTitle'),
-                      entries: res.map((k, v) => MapEntry(k, v.value)),
-                      selectedByDefault: false,
-                      onlyOneSelectionAllowed: true,
-                    );
-                  },
-                );
-          if (selectedUrls != null && selectedUrls.isNotEmpty) {
-            var sourceName = res[selectedUrls[0]]?.key;
-            changeUserInput(
-              selectedUrls[0],
-              true,
-              false,
-              updateUrlInput: true,
-              overrideSource: sourceName,
-            );
-          }
-        }
-      } catch (e) {
-        if (!context.mounted) return;
-        showError(e, context);
-      } finally {
-        setState(() {
-          searching = false;
-        });
-      }
-    }
-
     Widget getHTMLSourceOverrideDropdown() => Column(
       children: [
         Row(
@@ -608,106 +480,6 @@ class AddAppPageState extends State<AddAppPage> {
         const SizedBox(height: 16),
       ],
     );
-
-    bool shouldShowSearchBar() =>
-        sourceProvider.sources.where((e) => e.canSearch).isNotEmpty &&
-        pickedSource == null &&
-        userInput.isEmpty;
-
-    Widget getSearchBarRow() {
-      final ColorScheme colorScheme = Theme.of(context).colorScheme;
-      final bool searchDisabled =
-          searchQuery.isEmpty || doingSomething;
-      final Widget trailingSearch = searching
-          ? SizedBox(
-              width: 48,
-              height: 48,
-              child: Center(
-                child: SizedBox(
-                  width: 22,
-                  height: 22,
-                  child: CircularProgressIndicator(
-                    strokeWidth: 2,
-                    color: colorScheme.primary,
-                  ),
-                ),
-              ),
-            )
-          : Material(
-              color: searchDisabled
-                  ? colorScheme.primary.withValues(alpha: 0.38)
-                  : colorScheme.primary,
-              shape: const CircleBorder(),
-              clipBehavior: Clip.antiAlias,
-              child: InkWell(
-                customBorder: const CircleBorder(),
-                onTap: searchDisabled
-                    ? null
-                    : () {
-                        _searchSomeSourcesFocusNode.unfocus();
-                        FocusManager.instance.primaryFocus?.unfocus();
-                        HapticFeedback.selectionClick();
-                        runSearch();
-                      },
-                child: SizedBox(
-                  width: 48,
-                  height: 48,
-                  child: Icon(
-                    Icons.search,
-                    color: colorScheme.onPrimary,
-                    size: 22,
-                  ),
-                ),
-              ),
-            );
-      return Row(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          Expanded(
-            child: TextField(
-              focusNode: _searchSomeSourcesFocusNode,
-              controller: _searchSomeSourcesController,
-              onChanged: (String text) {
-                setState(() {
-                  searchQuery = text.trim();
-                });
-              },
-              textInputAction: TextInputAction.search,
-              onSubmitted: (_) {
-                if (!(searchQuery.isEmpty || doingSomething)) {
-                  _searchSomeSourcesFocusNode.unfocus();
-                  HapticFeedback.selectionClick();
-                  runSearch();
-                }
-              },
-              decoration: InputDecoration(
-                hintText: tr('searchSomeSourcesLabel'),
-                isDense: true,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(30),
-                ),
-                enabledBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(30),
-                  borderSide: BorderSide(
-                    color: colorScheme.outline.withValues(alpha: 0.55),
-                  ),
-                ),
-                focusedBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(30),
-                  borderSide: BorderSide(color: colorScheme.primary, width: 2),
-                ),
-                contentPadding: const EdgeInsets.symmetric(
-                  horizontal: 16,
-                  vertical: 10,
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(width: 10),
-          trailingSearch,
-        ],
-      );
-    }
 
     Widget getAdditionalOptsCol() {
       final ColorScheme colorScheme = Theme.of(context).colorScheme;
@@ -933,58 +705,571 @@ class AddAppPageState extends State<AddAppPage> {
       ),
     );
 
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.surface,
-      bottomNavigationBar: pickedSource == null ? getSourcesListWidget() : null,
-      body: CustomScrollView(
-        shrinkWrap: true,
-        slivers: <Widget>[
-          CustomAppBar(title: tr('addApp')),
-          SliverToBoxAdapter(
-            child: Padding(
-              padding: const EdgeInsets.fromLTRB(12, 8, 12, 16),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: [
-                  getUrlInputRow(),
-                  const SizedBox(height: 16),
-                  if (pickedSource != null) getHTMLSourceOverrideDropdown(),
-                  if (shouldShowSearchBar()) getSearchBarRow(),
-                  if (pickedSource == null) ...[
-                    const SizedBox(height: 24),
-                    OutlinedButton.icon(
-                      onPressed: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                            builder: (_) => const BulkAddAppsPage(),
-                          ),
+    // ── Search mode widgets ────────────────────────────────────────────
+
+    final Set<String> searchSelectedStores =
+        _getSearchSelectedStores(settingsProvider);
+
+    // ── Inline search runner ───────────────────────────────────────────
+
+    runInlineSearch({
+      required AppsProvider appsProvider,
+      required SettingsProvider settingsProvider,
+    }) async {
+      _searchSomeSourcesFocusNode.unfocus();
+      FocusManager.instance.primaryFocus?.unfocus();
+      _searchResultFilterController.clear();
+      setState(() {
+        searching = true;
+        _searchHasSearched = false;
+        _searchResults = {};
+        _searchResultFilter = '';
+      });
+
+      try {
+        final List<String> selectedSourceNames = searchSelectedStores.toList();
+        if (selectedSourceNames.isEmpty) {
+          throw ObtainiumError(tr('noResults'));
+        }
+
+        List<MapEntry<String, Map<String, List<String>>>?> results =
+            (await Future.wait(
+          sourceProvider.sources
+              .where((e) => selectedSourceNames.contains(e.name))
+              .map((e) async {
+                try {
+                  Map<String, dynamic>? querySettings = {};
+                  if (e.includeAdditionalOptsInMainSearch) {
+                    querySettings = await showDialog<Map<String, dynamic>?>(
+                      context: context,
+                      builder: (BuildContext ctx) {
+                        return GeneratedFormModal(
+                          title: tr('searchX', args: [e.name]),
+                          items: [
+                            ...e.searchQuerySettingFormItems.map((e) => [e]),
+                            [
+                              GeneratedFormTextField(
+                                'url',
+                                label: e.hosts.isNotEmpty
+                                    ? tr('overrideSource')
+                                    : plural('url', 1).substring(2),
+                                autoCompleteOptions: [
+                                  ...(e.hosts.isNotEmpty ? [e.hosts[0]] : []),
+                                  ...appsProvider.apps.values
+                                      .where(
+                                        (a) =>
+                                            sourceProvider
+                                                .getSource(
+                                                  a.app.url,
+                                                  overrideSource:
+                                                      a.app.overrideSource,
+                                                )
+                                                .runtimeType ==
+                                            e.runtimeType,
+                                      )
+                                      .map((a) {
+                                        var uri = Uri.parse(a.app.url);
+                                        return '${uri.origin}${uri.path}';
+                                      }),
+                                ],
+                                defaultValue:
+                                    e.hosts.isNotEmpty ? e.hosts[0] : '',
+                                required: true,
+                              ),
+                            ],
+                          ],
                         );
                       },
-                      icon: const Icon(Icons.add_rounded),
-                      label: Text(tr('bulkAddApps')),
-                    ),
-                  ],
-                  if (pickedSource != null)
-                    FutureBuilder(
-                      builder: (ctx, val) {
-                        return val.data != null && val.data!.isNotEmpty
-                            ? Text(
-                                val.data!,
-                                style: Theme.of(context).textTheme.bodySmall,
-                              )
-                            : const SizedBox();
+                    );
+                    if (querySettings == null) {
+                      return null;
+                    }
+                  }
+                  return MapEntry(
+                    e.runtimeType.toString(),
+                    await e.search(searchQuery, querySettings: querySettings),
+                  );
+                } catch (err) {
+                  if (err is! CredsNeededError) {
+                    rethrow;
+                  } else {
+                    err.unexpected = true;
+                    if (!context.mounted) return null;
+                    showError(err, context);
+                    return null;
+                  }
+                }
+              }),
+        ))
+                .where((a) => a != null)
+                .toList();
+
+        // Interleave results from multiple sources
+        Map<String, MapEntry<String, List<String>>> res = {};
+        var si = 0;
+        var done = false;
+        while (!done) {
+          done = true;
+          for (var r in results) {
+            var sourceName = r!.key;
+            if (r.value.length > si) {
+              done = false;
+              var singleRes = r.value.entries.elementAt(si);
+              res[singleRes.key] = MapEntry(sourceName, singleRes.value);
+            }
+          }
+          si++;
+        }
+
+        if (!context.mounted) return;
+        setState(() {
+          _searchResults = res;
+          _searchHasSearched = true;
+        });
+      } catch (e) {
+        if (!context.mounted) return;
+        showError(e, context);
+      } finally {
+        if (mounted) {
+          setState(() {
+            searching = false;
+          });
+        }
+      }
+    }
+
+    // ── Search mode widgets ────────────────────────────────────────────
+
+    Widget getSearchBarRow() {
+      final ColorScheme colorScheme = Theme.of(context).colorScheme;
+      final bool searchDisabled = searchQuery.isEmpty || doingSomething;
+      final Widget trailingSearch = searching
+          ? SizedBox(
+              width: 48,
+              height: 48,
+              child: Center(
+                child: SizedBox(
+                  width: 22,
+                  height: 22,
+                  child: CircularProgressIndicator(
+                    strokeWidth: 2,
+                    color: colorScheme.primary,
+                  ),
+                ),
+              ),
+            )
+          : Material(
+              color: searchDisabled
+                  ? colorScheme.primary.withValues(alpha: 0.38)
+                  : colorScheme.primary,
+              shape: const CircleBorder(),
+              clipBehavior: Clip.antiAlias,
+              child: InkWell(
+                customBorder: const CircleBorder(),
+                onTap: searchDisabled
+                    ? null
+                    : () {
+                        _searchSomeSourcesFocusNode.unfocus();
+                        FocusManager.instance.primaryFocus?.unfocus();
+                        HapticFeedback.selectionClick();
+                        runInlineSearch(
+                          appsProvider: appsProvider,
+                          settingsProvider: settingsProvider,
+                        );
                       },
-                      future: pickedSource?.getSourceNote(),
-                    ),
-                  if (pickedSource != null) getAdditionalOptsCol(),
-                ],
+                child: SizedBox(
+                  width: 48,
+                  height: 48,
+                  child: Icon(
+                    Icons.search,
+                    color: colorScheme.onPrimary,
+                    size: 22,
+                  ),
+                ),
+              ),
+            );
+      return Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Expanded(
+            child: TextField(
+              focusNode: _searchSomeSourcesFocusNode,
+              controller: _searchSomeSourcesController,
+              onChanged: (String text) {
+                setState(() {
+                  searchQuery = text.trim();
+                });
+              },
+              textInputAction: TextInputAction.search,
+              onSubmitted: (_) {
+                if (!(searchQuery.isEmpty || doingSomething)) {
+                  _searchSomeSourcesFocusNode.unfocus();
+                  HapticFeedback.selectionClick();
+                  runInlineSearch(
+                    appsProvider: appsProvider,
+                    settingsProvider: settingsProvider,
+                  );
+                }
+              },
+              decoration: InputDecoration(
+                hintText: tr('searchSomeSourcesLabel'),
+                isDense: true,
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(30),
+                ),
+                enabledBorder: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(30),
+                  borderSide: BorderSide(
+                    color: Theme.of(context).colorScheme.outline.withValues(alpha: 0.55),
+                  ),
+                ),
+                focusedBorder: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(30),
+                  borderSide: BorderSide(
+                    color: Theme.of(context).colorScheme.primary,
+                    width: 2,
+                  ),
+                ),
+                contentPadding: const EdgeInsets.symmetric(
+                  horizontal: 16,
+                  vertical: 10,
+                ),
               ),
             ),
           ),
+          const SizedBox(width: 10),
+          trailingSearch,
+        ],
+      );
+    }
+
+    Widget getSearchStoreChips() {
+      final searchableSources =
+          sourceProvider.sources.where((e) => e.canSearch).toList();
+      if (searchableSources.isEmpty) return const SizedBox.shrink();
+      return Wrap(
+        spacing: 8,
+        runSpacing: 4,
+        children: searchableSources.map((source) {
+          final selected = searchSelectedStores.contains(source.name);
+          return FilterChip(
+            label: Text(source.name),
+            selected: selected,
+            onSelected: (value) {
+              setState(() {
+                if (value) {
+                  searchSelectedStores.add(source.name);
+                } else {
+                  searchSelectedStores.remove(source.name);
+                }
+                // Persist the updated deselection list
+                settingsProvider.searchDeselected = searchableSources
+                    .map((s) => s.name)
+                    .where((n) => !searchSelectedStores.contains(n))
+                    .toList();
+              });
+            },
+          );
+        }).toList(),
+      );
+    }
+
+    Widget getSearchResultsList() {
+      if (!_searchHasSearched && _searchResults.isEmpty) {
+        return const SizedBox.shrink();
+      }
+      if (_searchResults.isEmpty) {
+        return Padding(
+          padding: const EdgeInsets.only(top: 24),
+          child: Center(child: Text(tr('noResults'))),
+        );
+      }
+      // Apply filter if the user typed one.
+      final String filterQ = _searchResultFilter.trim().toLowerCase();
+      final entries = _searchResults.entries.where((e) {
+        if (filterQ.isEmpty) return true;
+        final title = e.key.toLowerCase();
+        final subtitle = e.value.value.join(' ').toLowerCase();
+        return title.contains(filterQ) || subtitle.contains(filterQ);
+      }).toList();
+
+      return Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: [
+          Padding(
+            padding: const EdgeInsets.only(top: 16, bottom: 8),
+            child: Text(
+              tr('addAppSearchResultsTitle'),
+              style: Theme.of(context).textTheme.titleSmall?.copyWith(
+                    color: Theme.of(context).colorScheme.primary,
+                    fontWeight: FontWeight.w600,
+                  ),
+            ),
+          ),
+          TextField(
+            controller: _searchResultFilterController,
+            onChanged: (v) => setState(() => _searchResultFilter = v),
+            decoration: InputDecoration(
+              hintText: tr('filter'),
+              prefixIcon: const Icon(Icons.filter_list_rounded, size: 20),
+              suffixIcon: _searchResultFilter.isNotEmpty
+                  ? IconButton(
+                      icon: const Icon(Icons.close, size: 18),
+                      onPressed: () {
+                        _searchResultFilterController.clear();
+                        setState(() => _searchResultFilter = '');
+                      },
+                    )
+                  : null,
+              isDense: true,
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(30),
+              ),
+              contentPadding: const EdgeInsets.symmetric(
+                horizontal: 16,
+                vertical: 10,
+              ),
+            ),
+          ),
+          if (entries.isEmpty)
+            Padding(
+              padding: const EdgeInsets.only(top: 24),
+              child: Center(child: Text(tr('noResults'))),
+            ),
+          const SizedBox(height: 8),
+          ...entries.map((entry) {
+            final displayTitle = entry.key;
+            final sourceName = entry.value.key;
+            final subtitleLines = entry.value.value;
+            return Card(
+              margin: const EdgeInsets.only(bottom: 6),
+              child: ListTile(
+                leading: SizedBox(
+                  width: 32,
+                  child: Center(
+                    child: _searchSourceIcon(sourceName),
+                  ),
+                ),
+                title: subtitleLines.isNotEmpty
+                    ? Text(
+                        subtitleLines.join(' · '),
+                        maxLines: 2,
+                        overflow: TextOverflow.ellipsis,
+                      )
+                    : Text(
+                        displayTitle,
+                        maxLines: 2,
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                subtitle: subtitleLines.isNotEmpty
+                    ? Text(
+                        displayTitle,
+                        maxLines: 1,
+                        overflow: TextOverflow.ellipsis,
+                        style: Theme.of(context).textTheme.bodySmall,
+                      )
+                    : null,
+                trailing: const Icon(Icons.arrow_forward_ios_rounded, size: 16),
+                onTap: () {
+                  // Fill URL mode with selected result and switch to it
+                  changeUserInput(
+                    displayTitle,
+                    true,
+                    false,
+                    updateUrlInput: true,
+                    overrideSource: sourceName,
+                  );
+                  setState(() => _mode = _AddMode.byUrl);
+                },
+              ),
+            );
+          }),
+        ],
+      );
+    }
+
+    // ── Mode selector ──────────────────────────────────────────────────
+
+    Widget buildModeSelector() {
+      return Padding(
+        padding: const EdgeInsets.fromLTRB(12, 12, 12, 0),
+        child: SegmentedButton<_AddMode>(
+          segments: [
+            ButtonSegment(
+              value: _AddMode.byUrl,
+              label: Text(tr('addByUrl')),
+              icon: const Icon(Icons.link_rounded),
+            ),
+            ButtonSegment(
+              value: _AddMode.search,
+              label: Text(tr('addBySearch')),
+              icon: const Icon(Icons.search_rounded),
+            ),
+            ButtonSegment(
+              value: _AddMode.fromDevice,
+              label: Text(tr('addFromDevice')),
+              icon: const Icon(Icons.phone_android_rounded),
+            ),
+          ],
+          selected: {_mode},
+          onSelectionChanged: (v) => setState(() => _mode = v.first),
+          style: const ButtonStyle(
+            visualDensity: VisualDensity.compact,
+          ),
+        ),
+      );
+    }
+
+    // ── Layout ─────────────────────────────────────────────────────────
+
+    // Device mode uses a plain Column so the BulkAddWidget always gets a clean
+    // bounded height via Expanded, with no outer CustomScrollView that could
+    // steal scroll gestures or push content off-screen.
+    if (_mode == _AddMode.fromDevice) {
+      return Scaffold(
+        backgroundColor: Theme.of(context).colorScheme.surface,
+        appBar: AppBar(
+          title: Text(tr('addApp')),
+          automaticallyImplyLeading: false,
+        ),
+        body: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            buildModeSelector(),
+            Expanded(
+              child: BulkAddWidget(
+                key: _bulkWidgetKey,
+                onComplete: () => setState(() => _mode = _AddMode.byUrl),
+              ),
+            ),
+          ],
+        ),
+      );
+    }
+
+    return Scaffold(
+      backgroundColor: Theme.of(context).colorScheme.surface,
+      // Show supported-sources footer only for URL mode when no source is detected
+      bottomNavigationBar: (_mode == _AddMode.byUrl && pickedSource == null)
+          ? getSourcesListWidget()
+          : null,
+      body: CustomScrollView(
+        slivers: <Widget>[
+          CustomAppBar(title: tr('addApp')),
+          // Mode selector pinned just below the app bar
+          SliverPersistentHeader(
+            pinned: false,
+            delegate: _PaddedWidgetDelegate(
+              child: buildModeSelector(),
+              height: 60,
+            ),
+          ),
+          SliverToBoxAdapter(
+              child: Padding(
+                padding: const EdgeInsets.fromLTRB(12, 8, 12, 16),
+                child: AnimatedSwitcher(
+                  duration: const Duration(milliseconds: 180),
+                  child: KeyedSubtree(
+                    key: ValueKey(_mode),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.stretch,
+                      children: [
+                        // ── By URL ─────────────────────────────────────
+                        if (_mode == _AddMode.byUrl) ...[
+                          const SizedBox(height: 8),
+                          getUrlInputRow(),
+                          const SizedBox(height: 16),
+                          if (pickedSource != null)
+                            getHTMLSourceOverrideDropdown(),
+                          if (pickedSource != null)
+                            FutureBuilder(
+                              builder: (ctx, val) {
+                                return val.data != null && val.data!.isNotEmpty
+                                    ? Text(
+                                        val.data!,
+                                        style: Theme.of(context).textTheme.bodySmall,
+                                      )
+                                    : const SizedBox();
+                              },
+                              future: pickedSource?.getSourceNote(),
+                            ),
+                          if (pickedSource != null) getAdditionalOptsCol(),
+                        ],
+
+                        // ── Search ─────────────────────────────────────
+                        if (_mode == _AddMode.search) ...[
+                          const SizedBox(height: 8),
+                          getSearchBarRow(),
+                          const SizedBox(height: 12),
+                          Text(
+                            tr('storesToSearch'),
+                            style: Theme.of(context).textTheme.labelMedium?.copyWith(
+                                  color: Theme.of(context).colorScheme.onSurfaceVariant,
+                                ),
+                          ),
+                          const SizedBox(height: 6),
+                          getSearchStoreChips(),
+                          getSearchResultsList(),
+                        ],
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ),
         ],
       ),
     );
   }
+
+  /// Small icon to indicate which source a search result came from.
+  Widget _searchSourceIcon(String sourceName) {
+    final assetMap = {
+      'APKMirror': 'assets/graphics/ic_apkmirror.png',
+      'APKPure': 'assets/graphics/ic_apkpure.png',
+      'F-Droid': 'assets/graphics/ic_fdroid.png',
+      'GitHub': 'assets/graphics/ic_github.png',
+    };
+    final asset = assetMap[sourceName];
+    if (asset != null) {
+      return Image.asset(
+        asset,
+        width: 20,
+        height: 20,
+        fit: BoxFit.contain,
+        filterQuality: FilterQuality.medium,
+      );
+    }
+    return const Icon(Icons.store_rounded, size: 20);
+  }
+}
+
+/// Minimal [SliverPersistentHeaderDelegate] for a fixed-height widget.
+class _PaddedWidgetDelegate extends SliverPersistentHeaderDelegate {
+  final Widget child;
+  final double height;
+
+  const _PaddedWidgetDelegate({required this.child, required this.height});
+
+  @override
+  double get minExtent => height;
+
+  @override
+  double get maxExtent => height;
+
+  @override
+  Widget build(
+    BuildContext context,
+    double shrinkOffset,
+    bool overlapsContent,
+  ) {
+    return Material(
+      color: Theme.of(context).colorScheme.surface,
+      child: child,
+    );
+  }
+
+  @override
+  bool shouldRebuild(_PaddedWidgetDelegate oldDelegate) =>
+      oldDelegate.child != child || oldDelegate.height != height;
 }
